requirements.txt
accelerate==0.22.0
annotated-types==0.5.0
chromadb==0.3.29
elasticsearch
email-validator==2.0.0.post2
extra-streamlit-components==0.1.60
faiss-cpu==1.7.4
flask==2.3.3
httptools==0.6.0
httpx==0.25.0
install==1.3.5
langchain==0.0.281
openai==0.28.0
orjson==3.9.7
python-dotenv==1.0.0
python-multipart==0.0.6
rannet==0.3.1
st-annotated-text==4.0.1
protobuf==4.24.2
stanford-openie
tensorflow==2.13.0
transformers==4.33.0
ujson==5.8.0
unicorn==2.0.1.post1
unstructured==0.10.14
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
starlette==0.20.4
streamlit==1.26.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
nltk==3.8.1
langchain==0.0.281
langcodes==3.3.0
langml==0.4.2
langsmith==0.0.33
tiktoken==0.5.2